feat: add new compile script
Add compile script written in python that takes in a list of arguments
and can be configured using a json file.
For each schema we can define a base path and an output directory. If
the file is in the top level of the schemas dir base path can be
ommitted.

It is possible to define compile presets in the config this removes the
need for having to create multiple scripts.

We can either supply a list of schemas to compile, a preset or nothing
to compile all schemas defined in the schemas array. Required input is
one or more languages for which the compiler to run.

Information can be obtained by running ./capnp_compile.py --help.