feat: add additional package for schemas with stubs for type checking also add small pycapnp stubs
The package with stubs can be imported like so
from zalfmas_capnp_schemas_with_stubs import common_capnp

it is a drop in replacement for the old syntax
from zalfmas_capnp_schemas import common_capnp but now everything should
be type checked.
Also the capnproto package has now stubs installed for build it methods.

Expect some rough edges and maybe imprecise or wrong types but actually
it should be fairly robust as of now.

Knows issue is then using the _context parameter it has to have a
default value of None and we loose type checking of other function
parameters.