feat: add additional package for schemas with stubs for type checking… (#14)
* feat: add additional package for schemas with stubs for type checking also add small pycapnp stubs

The package with stubs can be imported like so
from zalfmas_capnp_schemas_with_stubs import common_capnp

it is a drop in replacement for the old syntax
from zalfmas_capnp_schemas import common_capnp but now everything should
be type checked.
Also the capnproto package has now stubs installed for build it methods.

Expect some rough edges and maybe imprecise or wrong types but actually
it should be fairly robust as of now.

Knows issue is then using the _context parameter it has to have a
default value of None and we loose type checking of other function
parameters.

* fix: include local package the proper way

* feat: add more capnproto type stubs

* feat: allow for dict assignment

* fix: add schemas_dir export for full compatibility

* chore: rewrite __init__.py of legacy packaging

* fix: proper return type of interface methods

* feat: support context results typing also support _new_client() method

* feat: use named tuple as server return value

* feat: export the named tuples so they can be instantiated